pyproject.toml
[project]
name = "nova"
version = "0.1.0"
description = "AI Research Assistant - A configurable CLI chatbot with chat history and memory"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "httpx>=0.25.0",
    "openai>=1.97.1",
    "anthropic>=0.60.0",
    "ollama>=0.5.1",
]

[project.scripts]
nova = "nova.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=1.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "asyncio: marks tests as async (deselect with '-m \"not asyncio\"')",
]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=nova",
    "--cov-report=term-missing",
    "--cov-report=html",
]

